Clean test setup, part I
Summary:
Our test setup is very arbitrary at the moment.
- We create files at arbitrary location.
- We invoke `Service.Parser.parse_source_files` without consistently taking care of all the shared memory cleanups after we are done.
- We pay no attention to the consistency of `Configuration`. Mock configurations are created at-will even though some of their settings contradict with each other.
- We use `File.Handle.create` and `Reference.create` to attach arbitrary handles/qualifiers to files even though the filesystem setup is not properly matched up.
- In most cases we skip setting up `ModuleTracker` in the test, despite the fact that `ModuleTracker` is used in production now for end-to-end checks and is specifically designed to handle the previous bullet.

These issues are big pain points in test maintenance, especially when I try to expand the `ModuleTracker` integration.

This diff introduces a unified API (i.e. `Test.ScratchProject.setup`) to simply the setup of the filesystem, the `Configuration`, and the `ModuleTracker`, and connect it to a number of tests. The philosophy there is that `Configuration`, `ModuleTracker`, and `Service.Parser` are all filesystem-state dependent modules, hence we should unify their creation logic and avoid mocking them to minimize the inconsistencies. See in-line comment in the diff for more details.

Reviewed By: dkgi

Differential Revision: D16207569

fbshipit-source-id: 1ac78480fc14213f8d353155669730fa2d83215f