Clean test setup, part III
Summary:
Our test setup is very arbitrary at the moment.
- We create files at arbitrary location.
- We invoke `Service.Parser.parse_source_files` without consistently taking care of all the shared memory cleanups after we are done.
- We pay no attention to the consistency of `Configuration`. Mock configurations are created at-will even though some of their settings contradict with each other.
- We use `File.Handle.create` and `Reference.create` to attach arbitrary handles/qualifiers to files even though the filesystem setup is not properly matched up.
- In most cases we skip setting up `ModuleTracker` in the test, despite the fact that `ModuleTracker` is used in production now for end-to-end checks and is specifically designed to handle the previous bullet.

These issues are big pain points in test maintenance, especially when I try to expand the `ModuleTracker` integration.

This diff is a follow-up to D16207569, where we try to clean up `serverTest`. The biggest issue with this set of tests is that we mock server state: the parser output, the environment output, and the type errors can all get created from nowhere, with no guarantee that they will be synced up. To fix this issue, we introduce a unified API (i.e.  `CommandTest.ScratchServer.start`) to start a server with a proper setup of `Configuration.Analysis`, `Configuration.Server`, and `Server.State`. Similar to the way we wrote `Test.ScratchProject.setup`, the philosophy there is that we want everything in the constructed state (except for the connections) to be derivable from a given filesystem state and we don't want the client to mock anything there.

It turns out that connecting `ScratchServer` to all the tests ends up with almost a rewrite of `serverTest`!

Reviewed By: dkgi

Differential Revision: D16213843

fbshipit-source-id: 81fe3a4ed47a34e0ffa1b5c0ac8395cbbb1dd31a