Restrict writes to `Ast.SharedMemory.Modules` within the parsing phase
Summary:
Let's do some real work here.

We want cleaner phase separation for better code maintainability and easier dependency management: each phase should be side-effect free from the view of other phases. It is not desirable to have, say, the parsing phase and the environment building phase to write to the same memory region.

Unfortunately, today the `Ast.SharedMemory.Modules` table is such piece of memory that is write-shared by both the parsing and the environment building phase. Logically speaking, this table will be cleared out after the parsing phase so it is not *really* shared. Nevertheless this gets into the way of my grand scheme: hiding `Ast.SharedMemory` behind an immutable data structure to subsequent phases. So let's clean it up.

What I did here is to add a `Modules` table to `EnvironmentSharedMemory`, and move all reference (except those that are located in the parser and preprocessor) of `Ast.SharedMemory.Modules` to point to the environment shared memory.

From now on, `Ast.SharedMemory.Modules` is going to be a parser-local legacy table waiting to get cleaned up when wildcard import resolution can be taken out from the parsing phase someday.

Reviewed By: dkgi

Differential Revision: D16224368

fbshipit-source-id: 6d91aa3019db41ecf6d9e25abb1e250eb06985dc