Batch lookup generation for queries
Summary:
Refactor lookup cache logic to ping the cache, generate lookups, and report failures on batches of files rather than only on one at a time. The old implementation allowed for file batches to be looked up sequentially, but we eventually want to support re-typechecking groups of files so we need to start handling them in batches.

 --- Bigger picture:

The ultimate goal here is to cut the reliance on store type check resolution when you have a running server and want to query types for file(s). We want a couple possible use cases and behaviors:

1. You've started the server with --store-type-check-resolution, probably because you're doing a codemod or something that anticipates requiring lookups for the majority of the codebase in large batches.
   - Allow batched lookups and use the stored resolution to generate them.
2. You've started a normal server and just want to ping it for types in a few files, probably for most other use-cases of type lookups including linting
   - Re-typecheck any necessary files in order to generate the lookup tables.
   - Both have some handling for requests on files that don't exist, as well as support for caching repeated lookups on files that haven't changed.

To support this on batches of files at once, we'd like to take all the files, check the cache for hits, separate out cache misses and if we've stored the type check resolution already, just generate the lookup tables from that resolution. If we haven't, then the cache misses should be batched and re-typechecked just like an incremental request where a few files have changed, while storing the resolutions for those files. We can generate the lookup tables, cache them, clear the resolution, and provide the lookup info.

Reviewed By: dkgi

Differential Revision: D16204750

fbshipit-source-id: 8f6900555680fea099178dd74f3da1d7890db943