Ignore all unsupported nested defines
Summary:
It is almost never the case we want to statically analyze a `def` inside if (except for processing version specific code, which is handled elsewhere).

The current accept-all strategy is particularly problematic for pysa, which cannot understand `def` inside if. However, the current implementation will pass those `def` to pysa and pysa will not filter them away and crash later due to invariant violation.

We want this recursive define visitor just for potential nested functions inside classes or functions, but never for if. Therefore, I make it always accepting class children but always reject visiting all other children. (In `MakeStatementVisitor` in `ast/visit.ml`, most statements are already excluded from visiting, and this change will additional exclude `With, For, If, While, Try`, which doesn't make sense for `def`s to live under them.)

Reviewed By: sinancepel

Differential Revision: D16268588

fbshipit-source-id: 02caa2b0723a586b40afafd395930e36526370e3