Connect `ModuleTracker` to incremental check
Summary: Now that all shared memory table key types are switched to `Reference.t`, and relevant tests are set up with `ModuleTracker`, connecting `ModuleTracker` to incremental check is relatively easy: instead of computing the handle change set in a two-pass, ad-hoc fashion, we directly invoke `ModuleTracker.update` (introduced in D15953191 and D15958470) to compute the module change set, and tweak the subsequent logic so that they are based on qualifiers instead of handles. It's a pretty straightforward refactor.

Reviewed By: dkgi

Differential Revision: D16234741

fbshipit-source-id: 746aac955a831fff38be3a2bd91b5cd2df46dbc1