ignore values assigned to in `__init__` when generating globals
Summary:
Due to Python's weird semantics with class variables and instance variables, we don't have a great way of detecting when a variable declared on a class toplevel is meant to be an annotation/declaration for a instance variable vs. actually a class variable (technically, you can annotate something as typing.ClassVar[T], but this isn't used often in practice, and misses unannotated variables completely).

Use the heuristic of "if self.x is assigned in `__init__`, `x` probably isn't a class variable".

Reviewed By: dkgi

Differential Revision: D16274022

fbshipit-source-id: 06ba6ffce483bd2cc144615b425c45f10ce37890