Move `SymlinksToPaths` into server state
Summary:
At the moment, the `SymlinksToPaths` table is only useful for our editor integration. This is a piece of info that matters to the server but not to serverless use cases.

This diff moves the info from `Ast.SharedMemory` into `Server.State`, and moves the table construction logic from `Service.Parser` into `Server.Operations`. I also changed the name of this table: the original name `SymlinksToPaths`  is very misleading because we are not actually mapping symlinks to the paths they pointed to, but the other way around.

Server start will be slightly delayed since we lose parallelized construction. But we can always add it back if this turns out to be a bottleneck.

Reviewed By: sinancepel

Differential Revision: D16284967

fbshipit-source-id: 840f132c6214fcb47db581cf8eb4855b44b34c0d