propagate global models when analyzing calls
Summary:
We were losing taint information on cases such as `MyClass.sink_list.append(source)` which should be caught as a flow of dangerous information.

The change in forwardAnalysis should make it so that we catch direct flows. The change in backwardAnalysis should propagate side effects, as caught in the indirect case.

The global models have been refactored to only expose the sink API, as that's all that's being used at the moment.

Reviewed By: fahndrich

Differential Revision: D16230349

fbshipit-source-id: 9fe44d4b96fcbf7927f0448f1ad7e144d84c9292