Refactor visit to add node traversal as base
Summary:
We're starting to duplicate searches through ASTs for nodes and locations in multiple places, including lookups, testing, location collection, etc. This is because expressions and statements aren't the only kinds of nodes that exist in the AST.

Working on writing up integration testing for nodes and locations, comparing a generated AST with a JSON test suite - didn't want rewrite this special casing yet again, so first working on supporting visits on Nodes in order to refactor away the existing special casing and use this for the integration testing, too.

Reviewed By: sinancepel

Differential Revision: D16287078

fbshipit-source-id: 1e776d6abfbb238dca4f7f03cadaa237490d578b