Implementing propagation functions for variadic classes
Summary:
The trickiest part of implementing user-accessible variadic classes is the "propagation" functions.  These are:
* `ClassHierarchy.instantiate_successors_parameters`
* `ClassHierarchy.instantiate_predecessors_parameters`
* `TypeOrder.instantiate_protocol_parameters`

These are the things that actually do the manipulation of parametric types, and establish the semantics of their relationships.  Because of the encapsulation work I've done over the past week, implementing the new behavior ended up being relatively clean.

Overall the approach I used was to achieve this was the following:
* Change the interface of ClassHierarchy to use `Type.OrderedTypes` in all places they used to use `Type.t list`, including `connect` and the propagation methods.
* Change the interface of ProtocolAssumptions and instantiate_protocol_parameters to take and return `OrderedTypes` respectively.
* Implement the inner logic of these functions to use the pre-existing ordered type solution logic.
* Fix up all the tests that used to expect the old interfaces.

Reviewed By: dkgi

Differential Revision: D16289571

fbshipit-source-id: 52c79484c0c8fbb1ee15cf2c5b52708387148797