Tighten up interface for ClassHierarchy.variables
Summary:
`ClassHierarchy.variables` is the interface for determining the parameters that can be attached to a given primitive class.  This should only ever return a list of type variables, rather than a list of types.
Rather than making downstream consumers clean up this list, it makes sense to keep that logic within classHierarchy.
Furthermore, this makes widening that interface to allow for variadics simpler.

Reviewed By: sinancepel

Differential Revision: D16270378

fbshipit-source-id: e4bd34a4bcc43daddafa1dec611ac51383b0d3b9