add support to track enumeration fields and constants
Summary:
The `first-index` feature is often used to filter enums. Our taint analysis was losing the name of the enum being accessed.

Go with the heuristic of only keeping the last element of all upper case simple names, as this will also automatically handle enums & constants that are more complicated than strings or integers.

Reviewed By: fahndrich

Differential Revision: D16347719

fbshipit-source-id: 7df701e133994dd558164a4576b71df0a5b2db21