Change the key type of `Ast.SharedMemory.Handles` from `int` to `Reference.t` [2/2]
Summary:
Goal
====
The initial motivation of introducing the `Handles` table is to eliminate the need of storing path string in every `Location` object, which can be quite expensive since we have tons of `Location`s in the AST.

Since `Reference.t` typically occupies more memory than `int`, moving from the latter to the former seems questionable at first glance as we'll only increase the memory pressure. However, if we realize that if we use `Reference.t` as the key, then all information stored in this `Handles` table would already be covered by `ModuleTracker`. Therefore we are able to completely drop this table once the migration is complete.

There's another reason why I want to switch to `ModuleTracker` for this use case. Currently, the error locations we reported for `BaseError.t` relies on `Handles.get` to be instantiated. `Handles` only stores file handle info, i.e. the relative paths of the source file. It does not know anything about which search path the source file lives under. This is fine for us in the past because we'd only type-check source files under local root. As we now start to type check stubs files, our previous assumption of "every source file lives under local root" is broken, leading to confusing error reports (see distillery error change in D15966252 or D16240594 for an example). To solve this problem we'd want to put more info in the `Handles` if we were planning to keep this, at which point just switching to `ModuleTracker` becomes a more enticing option.

This Diff
======

We make the switch on the `Handles` key type as well as the type signature of `Location.Reference`.

Reviewed By: dkgi

Differential Revision: D16341951

fbshipit-source-id: 7dbb83d8a9508a758962a52e4b2c082d3b8022e4