Make protocols play nice with metaclasses
Summary:
In order to get the attributes of "Type[X]", we don't actually want to look up the class definition of class, but rather do so for X and then get the class attributes off that definition.  Luckily that logic already exists in `Annotated.Class.resolve_class`.

There is probably a case to be made that Resolution.class_definition should be removed to avoid this in the future.

Reviewed By: grievejia

Differential Revision: D16350334

fbshipit-source-id: 4883049ad5cb4bbf38298b3392c4300aaa827a37