Add lookup location integration testing
Summary:
We'd like to have a suite of location tests that are shareable with libCST, to ensure that we don't regress and go out of sync. This is the reasoning for keeping the test cases separate and in JSON format (for now).

All we care about is that the list of expression/location pairs given in the test for a particular source is a subset of the expression/location pairs we know about when we parse that source - Pyre or LibCST may have additional locations we know about internally, but that's alright for the sake of these tests.

TODOs after this initial diff:

- Investigate YAML as an alternative to JSON - or a way to express the source code in a more readable way if it gets longer (multiline)
- Add fields for node_type for both pyre and libCST, and verify the pyre one
- Add fields for comments on the tests

Reviewed By: sinancepel

Differential Revision: D16343518

fbshipit-source-id: 8d186405456da1850dd87e984c7617a87640f715