Actually allowing variadics in parametrics
Summary:
This is the diff all of this has mostly been building up to, actually changing `Type.Parametric` to accept a `Type.OrderedTypes.t` instead of a a `Type.t list`.

Making this change was actually very light on actual logic changes due to all the prep work I already did, but ended up being very heavy in converting existing usages to use `Concrete` to indicate their use of the original kind of parametric.

Because of this, this diff is less scary than it looks, as most of its lines are just adding !s to indicate that a concrete list really is a concrete list in tests.

The actual logic work are the following:
* Including parametrics in the local_collect and local_replace functions.
* Establishing how these parametrics are converted to expressions.
* Including them in implicit generic base construction.

I tested most of this, but the heavy integration testing will have to wait until the next diff, which implements the parsing necessary for these to actually exist in the wild.

Reviewed By: sinancepel

Differential Revision: D16289613

fbshipit-source-id: 766cb48548f6a042717f0ebb514f6593801d5903