Setup builder cache read and write
Summary:
We want to use cached generated thrift files if the source does not change. The first step would be setup cache read and write.
This commit creates the cache data structure and its read/write method. We will always return a cache. In case a cache read fails, we return a cache that will invalidate everything.

Reviewed By: grievejia

Differential Revision: D16348649

fbshipit-source-id: adf71b7bb7d60983e4f83fee44ab9ff784f5cf71