Splitting GlobalResolution out from Resolution
Summary:
A major barrier to our fine-grain dependency work is the circular dependence between TypeCheck and Environment.  However, most of that circle is actually pretty immaterial.

It turns out that Resolution is actually composed of two parts that should be independent
1. The original use of resolution was storing the current scope of annotations.  This is the "local" component
2. Because of its omnipresence, folks have been tacking on what are essentially environment helper functions onto Resolution.  It turns out that environment building only actually uses these components, so they ought to be separated.

By this logic, I partitioned all of the global elements of Resolution off into GlobalResolution, which we will eventually make into the phase boundary data type between  environment building and type checking.  The next step will be to define typecheck as being based off of this global resolution, rather than the bare environment.

But for now, we just need to do this one massive mechanical replacement of most usages of Resolution with GlobalResolution.

Reviewed By: grievejia

Differential Revision: D16366474

fbshipit-source-id: 97b70cce1ae0845e9b3764d9dc9f22a385800fcc