Introduce thrift building cache to fast buck builder
Summary:
This commit connects work of previous commits together. We will now starting to use /tmp/pyre/buck_builder_cache/thrift-gen as the output of all generated thrift files, and create symlink to the output path when necessary.
We store the all the information that uniquely identifies a thrift library building and only rebuild if we cannot find it in cache.
Now thrift library building has to be divided into 3 stages as shown in the comments, because we can't simply establish symlink between folders since we might have nested thrift projects.

Reviewed By: dkgi

Differential Revision: D16381547

fbshipit-source-id: b5bb5af9969f16d452128bf3a3454052d1ddf7a8