Attach type based breadcrumbs to models
Summary:
We infer some "type-based-features" like `type:scalar` when we analyze functions based on argument and return types. This is useful for filtering.

For built-ins like `len`, etc, we are missing these as there is no body to analyze.

This diff augments declared models in .pysa files with type-based-features from the type environment. This way, one can declare explicit tito models for `len` and they will automatically get the features based on types.

  def len(__o: TaintInTaintOut): ...

will attach `type:scalar` to the tito flow.

Reviewed By: dkgi

Differential Revision: D16390451

fbshipit-source-id: 824c156ed2685807f232a8eb6ad9c785eaec48c5