Killing recursive infer
Summary:
Recursive infer is currently broken, and it's presence as the one analysis that mutates the global environment is making further refactors supporting consistency work harder.
We have decided to redo this functionality through actually writing stubs and re-parsing rather than directly mutating the environment.
For now, until that work is done, we're just killing this interface.

Reviewed By: shannonzhu

Differential Revision: D16421052

fbshipit-source-id: eb5ade2284a16b52bb5801add3a399b76faeeac8