Actually cache wheel downloads
Summary:
We didn't actually cache the downloaded wheel since I forget to remove the line
```
temporaryZipFile.delete();
```
when I first introduced the wheel cache.
This commit removes the line to actually cache the wheel. `temporaryZipFile` is also renamed to `downloadedZipFile` to avoid further confusion.

In this process, I find that the path to the downloaded zip is constructed incorrectly. The previous implementation might put the source in the working directory since we only keep the filename. But the problem is hidden since we will always delete the zip afterwards.

Reviewed By: sinancepel

Differential Revision: D16441891

fbshipit-source-id: bd0dcbabadde4c7594478437cca86c6248da6a42