Moving special classes out of Environment.Builder
Summary:
This diff removes the special classes from Environment.Builder in our quest to eliminate that submodule, on our way to sealing away handler overall.
These aren't as easy to eliminate as those we can directly inline into another builder (D16426754) or into a substitution map (D16426654).
Instead we need to add a helper method to add them all in to an arbitrary handler.

Reviewed By: grievejia

Differential Revision: D16433715

fbshipit-source-id: eecaa7b16488f7495e543d52a0d699810ed254b4