Stops unions containing Any from being treated as Any
Summary:
Unions (including Optionals) containing Any were being treated as Any, which led to return types like Optional[Any] being allowed for functions with integer or string return types. While Any is special and can be narrowed to any type, unions including Any are not so special. With this change, if a union contains a type that cannot be substituted for a function's return type then an error will be thrown.

The actual code change to accomplish this is small. Three existing tests required updates to their "expect files," but no test required logical changes to pass with this change. Four new tests were added to validate the expected behavior.

Changing this treatment of unions could still have implications not revealed by the tests, however.

Thanks to shannonzhu for her helpful ideas and pointers.

Note: the task originally mentions dealing with Top analogously, however this diff only deals with Any. It is unclear if unions containing Top should be collapsible or not, and naively doing so caused some additional test failures that seemed logical, although I did not pursue them.

Reviewed By: shannonzhu

Differential Revision: D16382520

fbshipit-source-id: 518c8bc970bebc94cfc5da8506d20851bbe14f03