Removing the only production use of Environment.Builder
Summary:
The only actual production use of `Analysis.Environment.Builder` was `Service.Environment.populate_shared_memory` which used that default in memory environment to populate the shared memory.  Now that we have removed everything from that default environment (except for the default `ClassHierarchy`), we can remove that call and only have to worry about dumping `ClassHierarchy.Builder.default` into shared memory, which is still kind of gross, but is much better than it once was.

There are much more usages in tests, which actually use the in process version of environment, but fixing those is gorier, and I'll split it into another diff.

Reviewed By: grievejia

Differential Revision: D16449223

fbshipit-source-id: 1f253c45685cb66ea921a7caf77b44c6d74dee46