Remove Environment.Builder from tests
Summary:
The last remnant of Environment.Builder was in many different tests.
By removing those usages, we can finally remove Environment.Builder, along with the corresponding `Environment.t`.
The only wrinkle here is that we move the table initialization into what was Environment.handler, rename it to in_process_handler, and add an optional argument to override the dependency for testing.

Reviewed By: grievejia

Differential Revision: D16450557

fbshipit-source-id: 6426ad634c7cae7101f39a368708c65d636dac63