Introducing `AstEnvironment`
Summary:
This is an attempt at implementing the new data structure to replace the global structure `Ast.SharedMemory`. Let's call it `AstEnvironment` for now, because I'm super creative when naming stuffs :P

Here are some desirable properties with this new set of APIs over the existing ones in `Ast.SharedMemory`:
- Access to global value is more restricted: there's no interface that directly lets one put data into or grab some data out of the shared memory tables.
Ideally, this should be achieved by hiding the `SharedMemory` modules from the mli file, and forcing every get/set interface to provide a reference to AstEnvironment.t first.
At the moment, due to the large amount of code referencing `Ast.SharedMemory.Sources` and `Ast.SharedMemory.Handles`, the hiding part cannot be carried out easily yet. Those tables need to stay exposed until we switch all existing `Ast.SharedMemory` reference to `AstEnvironment`. This diff is just the starting point where we provide the target to switch to.
- The underlying table implementation is not exposed. We should be able to freely switch between in-memory table and shared-memory table without changing how external code uses `AstEnvironment`.
In the code, this is achieved by making AstEnvironment.t an opaque type.
- We only want a small amount of ocaml code (i.e. the parser and maybe the incremental checker) have write access to the data. Every other pyre phase (e.g. environment building, type checking) should get a read-only view of the `AstEnvironment`.
In the code, this is achieved by defining a module (called `AstEnvironment.ReadOnly`) where a read-only subset of `AstEnvironment` interfaces are exposed. There's an API to convert `AstEnvironment` to `AstEnvironment.ReadOnly` but not vice versa, and it is expected that functions like `Environment.populate` and `TypeCheck.run` only works with `AstEnvironment.ReadOnly.t`, not `AstEnvironment.t`.

Reviewed By: dkgi

Differential Revision: D16454460

fbshipit-source-id: 0001b378e5b2677726304ccd2073c74454c3a182