Sealing away Environment.Handler
Summary:
Now that we have finally been able to kill Environment.Builder, we can go ahead and accomplish the main objective: sealing away Environment.Handler.
We want all of our accesses to the environment to go through a clean read/write interface, or a clean read only interface.  Now, we have that, with Analysis.Environment as the read/write interface and GlobalResolution as the read only interface.

In order to keep this diff simple I had to expose a few more things in the interface of Analysis.Environment, but there was already an audit of that interface coming, so adding a few more leaks for now is fine.

Another twist is that we needed to move the environment shared memory into analysis, in order to move the shared memory handler into Analysis.Environment.
Eventually we'll want to move that directly into Analysis.Environment.

The next step will be to do these same kinds of cleanups for Analysis.ClassHierarchy.Handler.

Reviewed By: dkgi

Differential Revision: D16459422

fbshipit-source-id: df20c81b40b46fe56ccd3ff4c310031aed8160ce