Removing mentions of environment shared memory from Service.Environment
Summary:
More cleanup, the only wrinkle here is the perf optimization to only prep the local changes for globalkeys, rather than doing a full transaction.
If we don't care about that cost, then we can remove that optional argument, but for now I'm trying to keep these as all no-ops.

Reviewed By: grievejia

Differential Revision: D16465210

fbshipit-source-id: 98dc0ab29911aa8a47ff7cb68257a6aad295a3dd