Adding a safe_cast in pyre that functions similar typing.cast but enforces subclass.
Summary: This adds safe_cast strucutres in the ML files, the pyre extension python, a new error type, unit tests, and a positive/negative integration test.  There are two current issues that should be addressed in this submit, even though it functions.  The first is that the return type for a literal passed into safe_cast returns typing.cast with the error.  The second is the error message can be better fleshed out by producing the value class for the user.

Reviewed By: shannonzhu

Differential Revision: D16188773

fbshipit-source-id: dcd31c4ca4a2bacd098437826873936680810734