Do not generate placeholder stubs in parallel
Summary:
Inside the stub generation we will add string to unsupported file set, which is not thread safe.
We don't usually need to generate a lot of stubs since the builder already supported most of the use cases. (Even if we need, the cost is still much smaller than thrift library generation)

Reviewed By: grievejia

Differential Revision: D16504705

fbshipit-source-id: a94a214a06a6938cf04624cedee1c0b5ba297f5a