Don't try to unmarshall string values
Summary: It turns out that hack parallel only marshalls value types that are not strings, so unmarshalling those kinds of values always fails.  The cleanest way that works with the type-unsafeness of marshall was to require an unmarshall function in each of the value types, and rely on the (forthcoming) integration test to enforce that people pick the right option.

Reviewed By: sinancepel

Differential Revision: D16503503

fbshipit-source-id: 24f6ab778c29149345727673c3eed6995ab51778