Supported nested thrift library
Summary:
This problem actually exists for a long time, but only when we introduced the cache it becomes visible enough.

Problem:
Suppose we have two thrift libraries that will be generated into `foo` and `foo/bar` respectively. If we are unlucky, we might generate the `foo/bar` before `foo`. However, in this way, during the `foo` generation, the thrift compiler will wipe away the entire `bar` folder inside `foo`.

Solution:
Instead of generating everything in one folder, we generate the code to the `THRIFT_CACHE_ROOT/[hash of the base module path of the target]`. Therefore, these files will never interfere again.

Reviewed By: grievejia

Differential Revision: D16504693

fbshipit-source-id: a846b6dd31077c4e28ad5200cf87b40dbda6bd1f