compute callables underlying @properties in CallResolution
Summary: We'd like to be able to analyze the body of properties for the purposes of taint analysis, which requires us to not be blind to the body of the functions. We were relying on trying to get the type of the direct property, which doesn't work in the general case. Use the attribute access logic to get the correct parent.

Reviewed By: fahndrich

Differential Revision: D16509430

fbshipit-source-id: d8ed5a103fd5389cd7b1ff2a813d5abda79f6cb0