Switch to `AstEnvironment` for `Service.Parser`
Summary:
We want the parser to create and returns an `AstEnvironment` so that we can pass it to the subsequent phases. More specifically, `Parser.parse_sources` takes an additional `AstEnvironment.t` argument, and `Parser.parse_all` returns an additional `AstEnvironment.t`.

The incremental check also needs to mutate `AstEnvironment`, so we need to put a mutable version of `AstEnvironment` into the server state. This unfortunate situation also have other consequences, such as `Check.check` needs to return a mutable version of `AstEnvironment` as well. We also have to update the tests accordingly due to these API changes.

Reviewed By: mrkmndz

Differential Revision: D16509509

fbshipit-source-id: cf75a817d254bf14a84ec33cf81d7a8bd0d6ec33