Pass only immutable local type information to nested defines
Summary:
We should not have any idea what the refined type of a global is when we access it inside a nested define.

For example:

```
GLOBAL: Union[int, str] = 1

def foo() -> int
   return GLOBAL   # this should be a problem!!
```

It doesn't matter that `foo` is defined right after the toplevel assignment of `GLOBAL = 1`. We have no idea when `foo` will be invoked, and `GLOBAL` could be anything (within limits of its given annotation) by then.

Trouble is, this gets more complicated when the outer scope is not the toplevel, since in the toplevel all assignments are globals that should be annotated. This also applies to locals that do not have a given annotation:

```
def outer() -> None:
   x = 1
   def inner() -> int:
       return x
   x = "string"
   inner()
```

In order to know what `x` is inside of the nested define, `x` needs to be given an annotation when it's assigned.

This diff solves only the first of these two problems - it strips globals of a local inferred type when passing a resolution on to a nested function.

RFC: The second is a little more invasive, and I wonder if we should be rethinking our inference of local variables assigned without an annotation. Note: if we chose to change our local variable mutability, it would mean we cannot do something like:

```
x = 1
x: str = "string"
```

Once we define a variable, its type cannot be overwritten / no local of the same name can be declared later.

The alternative, of course, is to cast all of these locals with no annotation to Any when creating the resolution for a nested define. They exist, but we have no idea what the type is, so treat it as say, an untyped parameter input.

Reviewed By: dkgi

Differential Revision: D16507647

fbshipit-source-id: 68ad782ce6f6f3de8af75547c979d5a85dcbf63f