Don't use the wrong variables for typing.Mapping in ClassHierarchy.Builder
Summary: `typing.Mapping` is covariant in its second parameter, and those are not the names of the variables.  Furthermore, it is unnecessary to add these variables, since they will be added when we parse in `typing.pyi`.  Adding these first is fine while we use the current strategy to merge in this in-process class hierarchy into shared memory, but a future refactor requires this to be correct or not added at all.

Reviewed By: dkgi

Differential Revision: D16548827

fbshipit-source-id: 71b0df75ba2ae43acaf127c3d41ffb897b09a344