make `pyre kill` also pkill the pyre client
Summary:
The interaction users want when the Pyre client is stuck is for `pyre kill` to restore things to a known good state. This works for the binary, but fails to consider the case when the client lock is taken while building large projects.

Kill the client name as well to make this work. Instead of using `pkill`, uses `pgrep` to avoid the race where `pkill` kills the calling process before it has a chance to kill the others.

Reviewed By: dkgi

Differential Revision: D16605533

fbshipit-source-id: ae6abd0d7c4b9aa37ae7ea89c0e83bff86fb96aa