support `ViaValueOf[parameter]]` annotations for sources, sinks and Tito
Summary:
Adds support for ViaValueOf annotations for sources, sinks, and tito. These will allow users to annotate taints with the arguments a function was called with.

The `ViaValueOf` syntax is introduced as a new kind of simple feature. We now support parsing models using this syntax, and compute features when applying calls with `ViaValueOf` annotations.

Reviewed By: fahndrich

Differential Revision: D16571750

fbshipit-source-id: ea7d6e89b14a4dbfc9b6984b3285ce357fabc480