Implementing Concatenation operator
Summary:
This diff implements a planned addition to our variadic type variable operators, concatenation.

For now, we are implementing a limited version of general concatenation.
Concatenate is defined as working on a single ListVariadic, or a single Map on a ListVariadic, concatenated with some non-zero number of unary types either coming before or after the "middle" member.

The reason for this restriction is the following:
```
def foo(t: Tuple[Concatenate[TsA, TsB]]) -> Tuple[Tuple[TsA], Tuple[TsB]]: ...

def bar(t: Tuple[int, str, bool]) -> None:
    reveal_type(foo(t)) # could reveal Tuple[Tuple[()], Tuple[int, str, bool]]  or
                        #              Tuple[Tuple[int], Tuple[str, bool]] or
                        #              Tuple[Tuple[int, str], Tuple[bool]] or
                        #              Tuple[Tuple[int, str, bool], Tuple[()]]
```
We could very well enumerate all of these possible constraints, but for the purposes we envision this being used for, it seems not that useful.

In order to maintain this guarantee in the type system, we had to encode a concatenation in kind of a funny way, as a separate head and tail which are typed as unaries, then a single middle member which can either be a variable or a mapped variable.

Even with these restrictions, we still run into situations where our constraint solving becomes tricky, as in `Concatenation[int, Ts] ≤ Concatenation[Ts, str]`.  We currently abdicate on all concat <-> concat / concat <-> map solves, and will maybe return if necessary (T48180915).

The next step is allowing `Generic[Concatenate[DType, ShapeTs]]` (T48185332).

Reviewed By: dkgi

Differential Revision: D16597092

fbshipit-source-id: 49d34e0c2d7e274eda2147698140dd2329c5c64a