refactor analysis to introduce awaitables at calls
Summary:
We were relying on a heuristic of "if the expression is a top-level expression or an assignment value, check if it's an awaitable".
The previous approach had the limitation that we had to, in general, introduce unawaited awaitables for each statement pattern. What happens in the runtime is that an awaitable gets introduced when
you call an asynchronous function, so make the analysis mimic this better.

As a side effect, we now error on awaitables nested in tuples, and fixes a bug where we were marking the wrong location as awaited for `await e`.

Reviewed By: grievejia

Differential Revision: D16650779

fbshipit-source-id: 461f42db0e42b068eee7975267c215dadc39b2c2