Introducing `Error.Instantiated`
Summary:
The ultimate goal here is to replace access to the global `Ast.SharedMemory.Handles` table with the new `AstEnvironment` interface.

This table was originally introduced in D8743535, where the idea is to save memory by storing long path strings with their int hash values. We need the table to convert those integers back to their path strings when locations need to be shown to the user.

Problem is, this conversion happens *during* type checking & post-processing, when `Error`s are created. But not all places where we create an `Error` gets handy access to an `AstEnvironment`. Example: creating unused ignore errors.

Moreover, it is desirable to have the type checker to be file-path-oblivious, and use module name (i.e. `Reference.t`) as a "universal pointer" to refer to locations instead. As explained in the summary section of D16341951, this makes it easy to change how paths are displayed to the user (e.g. include search path root to the error reports).

That brings us to the basic idea to this diff: We want to **delay** the instantiation of `Error.location` as much as possible, ideally only do it when displaying the errors and/or generating the JSON responses. There are many ways to achieve the goal. I took the approach of introducing an additional module `Error.Instantiated`, and move the original `Error.description` and `Error.to_json` to this new module.

Under the hood, the only difference between `Error.t` and `Error.Instantiated.t` is that the former holds a `Location.Reference.t` while the latter holds a `Location.Instantiated.t`. We introduce the function `Error.instantiate` as a bridge from `Error.t` to `Error.Instantiated.t`. The idea is that every time we want to do a `description` or `to_json` operation on an error, we have to instantiate the error first, while other operations are left unaffected. In type checking & post-processing, we can therefore create `Error.t` with reference locations without having to query `Ast.SharedMemory.Handles`.

The rest of the changes in this diff is just massaging the codebase so that it type checks.

Reviewed By: sinancepel

Differential Revision: D16648821

fbshipit-source-id: 49eb8a7f39c88be4366b0c02ca551d1fca5e292b